refactor : Header 컴포넌트 파일을 컨벤션에 맞게 이름 변경 및 폴더 구조화 및 임시 UI 수정
